Wait longer for add-on setup, reload install page if it times out
When the add-on is first installed from the "Install the Test Pilot
Add-on" button, we start an "Installing" animation. Once the add-on is
installed & set up, we try to detect that and move on to the next step.
If the add-on takes too long, we time out with an unhandled error.

This PR attempts to handle that error by reloading the page. If the
add-on was installed but just took awhile to set up, this should result
in the user seeing the experiments page. Otherwise, they'll see the
"Install the Test Pilot Add-on" button again

Fixes #1075