only build version.json if missing
due to technical limitations cloudops builds `dist/` on a separate machine from where it runs deploy.sh. the .git directory is not copied, and any files outside of dist/ have to be manually specified in order to be used.

so cloudops will add bin/build-version-json.sh to our build step, and this script should assume that if version.json exists, then it should not regenerate it.

this shouldn't break a workflow of committing locally and running bin/deploy.sh, because bin/deploy.sh moves version.json to dist/__version__, so bin/deploy.sh will regen version.json on each run.