Run ZAP baseline when deploying to dev environment
This change (hopefully;) allows us to run the ZAP baseline scan whenever
testpilot is deployed to the dev environment. This perform a 1 min
spider of testpilot and then checks for whatever we can check for on the
security checklist
https://wiki.mozilla.org/Security/CloudSec#Security_Checklist 
For more info about the ZAP baseline scan see
https://github.com/zaproxy/zaproxy/wiki/ZAP-Baseline-Scan
The current checks are actually defaulted and need to be tuned via a
config file - any suggestions as to where this could live in this repo?