Remove the en and de-DE locales
The en and de-DE locales were added in #2149 in order to fix #2017.  The
issue was related to the language negotiation done by l20n. It didn't
pick en-US nor de for some users.

In #2399 l20n was upgraded to v5.  It uses [fluent-langneg][1] under the
hood to conduct the language negotiation.  It now properly selects the
best available locale when the user's browser is configured to request
de-DE or en.

[1]: https://www.npmjs.com/package/fluent-langneg