Fix #2732 Fix localization by using document.querySelector('meta[name=availableLanguages]') instead of config.AVAILABLE_LOCALES
Apparently there are two files named config.js, and the one that is
available to the browser does not define AVAILABLE_LOCALES. The one
which does define AVAILABLE_LOCALES is not able to be required by the
browser.

Luckily, the meta tag availableLanguages is usable and always contains
the correct languages.