Stop including node polyfills in client-side JS
Figured out that [the seedrandom module optionally requires node's
crypto module][optional] for a feature that we don't seem to use.
However, that seems to prompt Webpack to pull everything in just in
case. Turning off the node-browser polyfills appears to save us ~309kb
in the vendor bundle without breaking anything.

[optional]: https://github.com/davidbau/seedrandom/blob/released/seedrandom.js#L237

Issue #2850