Fix #3096 Fix the broken email signup dialog by not checking cookies until componentDidMount.
Checking them in the constructor will always fail when rendering the page html at build time.

Honestly I have no idea why this fixes the problem, but it does and I had a hunch to try it because it is just cleaner to do things that require browser state in componentDidMount.