Remove Frame Scripts from addon
- fixes #3176

Within the webextension api we are no longer able to modify the
navigator object on mozilla.com or about:home, so snippets will no
longer be able to identify whether or not a user is in testpilot.
Additionally, we won't be providing information to the frontend
through the addon, this isn't an issue as we now use
mozAddonManager on the frontend.